+++ backend/libs/kvs/src/traits/symbol/expiration.rs
@@ -17,7 +17,7 @@ pub trait Expiration: Base + ChannelName {
     symbol: Arc<String>,
     dur: Duration,
   ) -> KVSResult<bool> {
-    let dur_mils = dur.as_millis() as usize;
+    let dur_mils = dur.as_millis() as i64;
     let mut cmd = self.__commands__();
     let channel_name = self.channel_name(exchange, symbol);
     if cmd
